refactor(persistence): Combine PromotionRepository & ArtifactRepository
