+++ keel-api/src/main/kotlin/com/netflix/spinnaker/config/DefaultConfiguration.kt
@@ -6,12 +6,10 @@ import com.netflix.spinnaker.filters.AuthenticatedRequestFilter
 import com.netflix.spinnaker.keel.api.ApiVersion
 import com.netflix.spinnaker.keel.persistence.ArtifactRepository
 import com.netflix.spinnaker.keel.persistence.DeliveryConfigRepository
-import com.netflix.spinnaker.keel.persistence.PromotionRepository
 import com.netflix.spinnaker.keel.persistence.ResourceRepository
 import com.netflix.spinnaker.keel.persistence.ResourceVersionTracker
 import com.netflix.spinnaker.keel.persistence.memory.InMemoryArtifactRepository
 import com.netflix.spinnaker.keel.persistence.memory.InMemoryDeliveryConfigRepository
-import com.netflix.spinnaker.keel.persistence.memory.InMemoryPromotionRepository
 import com.netflix.spinnaker.keel.persistence.memory.InMemoryResourceRepository
 import com.netflix.spinnaker.keel.persistence.memory.InMemoryResourceVersionTracker
 import com.netflix.spinnaker.keel.plugin.ResolvableResourceHandler
@@ -57,10 +55,6 @@ class DefaultConfiguration {
     artifactRepository: ArtifactRepository
   ): DeliveryConfigRepository =
     InMemoryDeliveryConfigRepository()
-
-  @Bean
-  @ConditionalOnMissingBean
-  fun promotionRepository(): PromotionRepository = InMemoryPromotionRepository()
 
   @Bean
   @ConditionalOnMissingBean(ResourceVersionTracker::class)
