+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/persistence/memory/InMemoryArtifactRepository.kt
@@ -2,13 +2,15 @@ package com.netflix.spinnaker.keel.persistence.memory
 
 import com.netflix.spinnaker.keel.api.ArtifactType
 import com.netflix.spinnaker.keel.api.DeliveryArtifact
+import com.netflix.spinnaker.keel.api.DeliveryConfig
 import com.netflix.spinnaker.keel.persistence.ArtifactAlreadyRegistered
 import com.netflix.spinnaker.keel.persistence.ArtifactRepository
 import com.netflix.spinnaker.keel.persistence.NoSuchArtifactException
 
 class InMemoryArtifactRepository : ArtifactRepository {
-  private val artifacts: MutableMap<DeliveryArtifact, MutableList<String>> =
-    mutableMapOf()
+  private val artifacts = mutableMapOf<DeliveryArtifact, MutableList<String>>()
+  private val approvedVersions = mutableMapOf<Triple<DeliveryArtifact, DeliveryConfig, String>, String>()
+  private val deployedVersions = mutableMapOf<Triple<DeliveryArtifact, DeliveryConfig, String>, MutableList<String>>()
 
   override fun register(artifact: DeliveryArtifact) {
     if (artifacts.containsKey(artifact)) {
@@ -38,7 +40,53 @@ class InMemoryArtifactRepository : ArtifactRepository {
   override fun versions(artifact: DeliveryArtifact): List<String> =
     artifacts[artifact] ?: throw NoSuchArtifactException(artifact)
 
+  override fun approveVersionFor(
+    deliveryConfig: DeliveryConfig,
+    artifact: DeliveryArtifact,
+    version: String,
+    targetEnvironment: String
+  ) {
+    val key = Triple(artifact, deliveryConfig, targetEnvironment)
+    approvedVersions[key] = version
+  }
+
+  override fun latestVersionApprovedIn(
+    deliveryConfig: DeliveryConfig,
+    artifact: DeliveryArtifact,
+    targetEnvironment: String
+  ): String? {
+    val key = Triple(artifact, deliveryConfig, targetEnvironment)
+    return approvedVersions[key]
+  }
+
+  override fun wasSuccessfullyDeployedTo(
+    deliveryConfig: DeliveryConfig,
+    artifact: DeliveryArtifact,
+    version: String,
+    targetEnvironment: String
+  ): Boolean {
+    val key = Triple(artifact, deliveryConfig, targetEnvironment)
+    return deployedVersions[key]?.contains(version) ?: false
+  }
+
+  override fun markAsSuccessfullyDeployedTo(
+    deliveryConfig: DeliveryConfig,
+    artifact: DeliveryArtifact,
+    version: String,
+    targetEnvironment: String
+  ) {
+    val key = Triple(artifact, deliveryConfig, targetEnvironment)
+    val list = deployedVersions[key]
+    if (list == null) {
+      deployedVersions[key] = mutableListOf(version)
+    } else {
+      list.add(version)
+    }
+  }
+
   fun dropAll() {
     artifacts.clear()
+    approvedVersions.clear()
+    deployedVersions.clear()
   }
 }
