+++ keel-core/src/test/kotlin/com/netflix/spinnaker/keel/persistence/memory/InMemoryArtifactRepositoryTests.kt
@@ -2,6 +2,10 @@ package com.netflix.spinnaker.keel.persistence.memory
 
 import com.netflix.spinnaker.keel.persistence.ArtifactRepositoryTests
 
-internal class InMemoryArtifactRepositoryTests : ArtifactRepositoryTests<InMemoryArtifactRepository>() {
+class InMemoryArtifactRepositoryTests : ArtifactRepositoryTests<InMemoryArtifactRepository>() {
   override fun factory() = InMemoryArtifactRepository()
+
+  override fun InMemoryArtifactRepository.flush() {
+    dropAll()
+  }
 }
