--- keel-core/src/test/kotlin/com/netflix/spinnaker/keel/persistence/memory/InMemoryPromotionRepositoryTests.kt
@@ -1,11 +1,0 @@
-package com.netflix.spinnaker.keel.persistence.memory
-
-import com.netflix.spinnaker.keel.persistence.PromotionRepositoryTests
-
-class InMemoryPromotionRepositoryTests : PromotionRepositoryTests<InMemoryPromotionRepository>() {
-  override fun factory() = InMemoryPromotionRepository()
-
-  override fun InMemoryPromotionRepository.flush() {
-    dropAll()
-  }
-}
